Rest of merging
