+++ backend/Data/DataContextFactory.cs
@@ -1,8 +1,8 @@
 using System.IO;
+using DotNetEnv;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Design;
 using Microsoft.Extensions.Configuration;
-using DotNetEnv;
 
 namespace backend.Data
 {
